Move pages into folders
