+++ src/routes/index.jsx
@@ -4,8 +4,8 @@ import { initAll } from 'govuk-frontend';
 
 import { useKeycloak } from '../utils/keycloak';
 import Layout from '../components/Layout';
-import TaskListPage from './TaskListPage';
-import TaskDetailsPage from './TaskDetailsPage';
+import TaskListPage from './TaskLists/TaskListPage';
+import TaskDetailsPage from './TaskDetails/TaskDetailsPage';
 import IssueTargetPage from './IssueTargetPage';
 
 const AppRouter = () => {
